buildInitialRoute is almost done
The concept of "super customer" has been introduced.
A "super customer" is the first customer assigned to a route. Super
customers are always the most distant from the depot.
Each supercustomer has its own neighbourhood of near customers. The
near customers are chosen based on the distance from the supercustomer.
Customers taken into a neighbourhood are marked with isTaken = true.