Some parts commented out to fix the cost issue
Some parts of the code (mainly on the tabu search package) were
commented out to try to fix the issue of cost "Infinity".
Still no luck. The whole logic of the tabu search should be rewritten
accordingly to this new approach taken for generating the initial route.
After some debugging, it seems that the initial route is correctly
generated.